Prefetch data to achieve better memory bandwidth utilization. With the new kernel token generation is ~48% faster for 33B model and ~14% faster for 7B model.  Tests for 65B model are welcome. 

|33B model| master | this PR |
|---             |---         | ----|
|GPU Read BW*| 167 GB/s | 275 GB/s|
|prompt 1| 117ms/tok | 79ms/tok|
|prompt 2 |121ms/tok | 82 ms/tok|


|7B model| master | this PR |
|---             |---         | ----|
|prompt 1| 25.7ms/tok | 22.5ms/tok|

GPUs of M1 Max / M2 Max can sustain 340 GB/s , can we reach that in llama.cpp? 😳 

\* GPU Read Bandwidth measured using Developer Tools comes with Xcode.
<img width="1099" alt="Screenshot 2023-07-12 at 01 24 05" src="https://github.com/ggerganov/llama.cpp/assets/61452103/05ae64b9-aefa-4ee1-b032-ec4c8d4a02b1">


\** Measured with the following command on M1 Max 32GB. Generation results are same between this PR and master branch.
`./main -m model_file -n 128 -c 512 -s 12 -ngl 1 --no-mmap` 
Prompt 1 “”
Prompt 2 “I believe the meaning of life is”