@lshzh-ww 2023-07-12
Please feel free to share results on M1/M2 Pro/Max/Ultra!