@ikawrakow 2023-07-12
>  Please feel free to share results on M1/M2 Pro/Max/Ultra!

M2 Max with 30-core GPU, 64 GB RAM:
* 7B is 21.5 ms/t on master, 21.7 ms/t with this PR
* 33B is 91.3 ms/t for master and this PR.

The above is for an empty prompt. If I use `-p "I believe the meaning of life is"` instead, I get ~92 ms/t on master and ~97 ms/t with this PR.  

Is there something special I need to do with the build to get the speed up?