@lshzh-ww 2023-07-12
> > Please feel free to share results on M1/M2 Pro/Max/Ultra!
> 
> Well, I had not checked out the PR branch. With the PR branch I get on M2 Max with 30-core GPU, 64 GB RAM:
> 
> * 7B is 21.5 ms/t on master, 21.0 ms/t with this PR
> * 33B is 91.3 ms/t for master and 77.4 ms/t with this PR.
> 
> The above is for an empty prompt. If I use `-p "I believe the meaning of life is"` instead, I get for 33B ~92 ms/t on master and ~84 ms/t with this PR.

Nice! I update the measurements with new prompt. Looks like this PR brings the M1 series to the same speed as M2 series. I guess M2 series have better memory access prediction so their speed are already good without this PR. Nonetheless for 65B model I guess we can see a larger performance improvement.