@lshzh-ww 2023-07-12
Would be nice to see results on M1 Ultra or M2 Ultra chips. Their two-die design may also suffer from cache miss.