@adrienbrault 2023-07-12
Here's my benchmark: MBP 14 Apple M2 Max 32GB 12C CPU, 30C GPU

| 13B model | master | this PR |
|-|-|-|
| prompt 1 1/2 | 46.97ms/tok | 36.70ms/tok |
| prompt 1 2/2 | 50.17ms/tok | 37.40ms/tok |



<details>
  <summary>Logs</summary>

```
17:53:48 in ~/Developer/ai/llama.cpp on  mps-q4_0-kernel 
✘130 ➜ ./main -m ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin -n 128 -c 512 -s 12 -ngl 1 --no-mmap -p ""
main: build = 821 (38ec9a2)
main: seed  = 12
llama.cpp: loading model from ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32001
llama_model_load_internal: n_ctx      = 512
llama_model_load_internal: n_embd     = 5120
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 40
llama_model_load_internal: n_layer    = 40
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: ftype      = 2 (mostly Q4_0)
llama_model_load_internal: n_ff       = 13824
llama_model_load_internal: model size = 13B
llama_model_load_internal: ggml ctx size = 6983.71 MB
llama_model_load_internal: mem required  = 9031.71 MB (+ 1608.00 MB per state)
llama_new_context_with_model: kv self size  =  400.00 MB
ggml_metal_init: allocating
ggml_metal_init: using MPS
ggml_metal_init: loading '/Volumes/Developer/ai/llama.cpp/ggml-metal.metal'
ggml_metal_init: loaded kernel_add                            0x102a04c70
ggml_metal_init: loaded kernel_mul                            0x102a054b0
ggml_metal_init: loaded kernel_mul_row                        0x102a05ae0
ggml_metal_init: loaded kernel_scale                          0x102a06000
ggml_metal_init: loaded kernel_silu                           0x102a06520
ggml_metal_init: loaded kernel_relu                           0x102a06a40
ggml_metal_init: loaded kernel_gelu                           0x102a06f60
ggml_metal_init: loaded kernel_soft_max                       0x102b06b60
ggml_metal_init: loaded kernel_diag_mask_inf                  0x102b07120
ggml_metal_init: loaded kernel_get_rows_f16                   0x102b077a0
ggml_metal_init: loaded kernel_get_rows_q4_0                  0x102b07e20
ggml_metal_init: loaded kernel_get_rows_q4_1                  0x102b08610
ggml_metal_init: loaded kernel_get_rows_q2_K                  0x13de074f0
ggml_metal_init: loaded kernel_get_rows_q3_K                  0x13de07c90
ggml_metal_init: loaded kernel_get_rows_q4_K                  0x13de08310
ggml_metal_init: loaded kernel_get_rows_q5_K                  0x13de08990
ggml_metal_init: loaded kernel_get_rows_q6_K                  0x13de09010
ggml_metal_init: loaded kernel_rms_norm                       0x13de096c0
ggml_metal_init: loaded kernel_norm                           0x13de09d70
ggml_metal_init: loaded kernel_mul_mat_f16_f32                0x13de0a860
ggml_metal_init: loaded kernel_mul_mat_q4_0_f32               0x13de0af40
ggml_metal_init: loaded kernel_mul_mat_vec_q4_0_f32           0x13de0b780
ggml_metal_init: loaded kernel_mul_mat_q4_1_f32               0x13de0bd40
ggml_metal_init: loaded kernel_mul_mat_q2_K_f32               0x13de0c5a0
ggml_metal_init: loaded kernel_mul_mat_q3_K_f32               0x102a07540
ggml_metal_init: loaded kernel_mul_mat_q4_K_f32               0x102a07d40
ggml_metal_init: loaded kernel_mul_mat_q5_K_f32               0x102a08420
ggml_metal_init: loaded kernel_mul_mat_q6_K_f32               0x102a08d00
ggml_metal_init: loaded kernel_rope                           0x102a095f0
ggml_metal_init: loaded kernel_alibi_f32                      0x102a09eb0
ggml_metal_init: loaded kernel_cpy_f32_f16                    0x102a0a740
ggml_metal_init: loaded kernel_cpy_f32_f32                    0x102a0b230
ggml_metal_init: loaded kernel_cpy_f16_f16                    0x102a0bac0
ggml_metal_init: recommendedMaxWorkingSetSize = 21845.34 MB
ggml_metal_init: hasUnifiedMemory             = true
ggml_metal_init: maxTransferRate              = built-in GPU
llama_new_context_with_model: max tensor size =    87.89 MB
ggml_metal_add_buffer: allocated 'data            ' buffer, size =  6983.72 MB, ( 6984.17 / 21845.34)
ggml_metal_add_buffer: allocated 'eval            ' buffer, size =  1024.00 MB, ( 8008.17 / 21845.34)
ggml_metal_add_buffer: allocated 'kv              ' buffer, size =   402.00 MB, ( 8410.17 / 21845.34)
ggml_metal_add_buffer: allocated 'scr0            ' buffer, size =   512.00 MB, ( 8922.17 / 21845.34)
ggml_metal_add_buffer: allocated 'scr1            ' buffer, size =   512.00 MB, ( 9434.17 / 21845.34)

system_info: n_threads = 8 / 12 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | VSX = 0 |
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 128, n_keep = 0



// RUN: %target-swift-ide-test -code-completion -source-filename %s -code-completion-token=GLOBAL_ENUM_1 > %t.globaleenum1.txt
// RUN: FileCheck --input-file=%t.globaleenum1.txt %s

// CHECK: CODE_ completionsForGlobalEnum1
// CHECK-NEXT: SYMBOL KIND
// CHECK-NEXT: BASE_ENUM=TypeName
// CHECK-NEXT: ◂────────────────────────
llama_print_timings:        load time =  4994.00 ms
llama_print_timings:      sample time =   180.25 ms /   128 runs   (    1.41 ms per token,   710.13 tokens per second)
llama_print_timings: prompt eval time =   375.17 ms /     2 tokens (  187.58 ms per token,     5.33 tokens per second)
llama_print_timings:        eval time =  4660.87 ms /   127 runs   (   36.70 ms per token,    27.25 tokens per second)
llama_print_timings:       total time =  5234.85 ms
ggml_metal_free: deallocating

17:54:12 in ~/Developer/ai/llama.cpp on  mps-q4_0-kernel   took 10.4s
➜ ./main -m ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin -n 128 -c 512 -s 12 -ngl 1 --no-mmap -p ""
main: build = 821 (38ec9a2)
main: seed  = 12
llama.cpp: loading model from ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32001
llama_model_load_internal: n_ctx      = 512
llama_model_load_internal: n_embd     = 5120
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 40
llama_model_load_internal: n_layer    = 40
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: ftype      = 2 (mostly Q4_0)
llama_model_load_internal: n_ff       = 13824
llama_model_load_internal: model size = 13B
llama_model_load_internal: ggml ctx size = 6983.71 MB
llama_model_load_internal: mem required  = 9031.71 MB (+ 1608.00 MB per state)
llama_new_context_with_model: kv self size  =  400.00 MB
ggml_metal_init: allocating
ggml_metal_init: using MPS
ggml_metal_init: loading '/Volumes/Developer/ai/llama.cpp/ggml-metal.metal'
ggml_metal_init: loaded kernel_add                            0x151706b40
ggml_metal_init: loaded kernel_mul                            0x151707260
ggml_metal_init: loaded kernel_mul_row                        0x151707890
ggml_metal_init: loaded kernel_scale                          0x151707db0
ggml_metal_init: loaded kernel_silu                           0x1517082d0
ggml_metal_init: loaded kernel_relu                           0x1517087f0
ggml_metal_init: loaded kernel_gelu                           0x151708d10
ggml_metal_init: loaded kernel_soft_max                       0x1517093c0
ggml_metal_init: loaded kernel_diag_mask_inf                  0x151709a20
ggml_metal_init: loaded kernel_get_rows_f16                   0x15170a0a0
ggml_metal_init: loaded kernel_get_rows_q4_0                  0x15170a720
ggml_metal_init: loaded kernel_get_rows_q4_1                  0x15170af10
ggml_metal_init: loaded kernel_get_rows_q2_K                  0x15170b590
ggml_metal_init: loaded kernel_get_rows_q3_K                  0x15170bc10
ggml_metal_init: loaded kernel_get_rows_q4_K                  0x15170c290
ggml_metal_init: loaded kernel_get_rows_q5_K                  0x15170c910
ggml_metal_init: loaded kernel_get_rows_q6_K                  0x15170cf90
ggml_metal_init: loaded kernel_rms_norm                       0x15170d640
ggml_metal_init: loaded kernel_norm                           0x15170dcf0
ggml_metal_init: loaded kernel_mul_mat_f16_f32                0x15170e6c0
ggml_metal_init: loaded kernel_mul_mat_q4_0_f32               0x15170eda0
ggml_metal_init: loaded kernel_mul_mat_vec_q4_0_f32           0x15170f440
ggml_metal_init: loaded kernel_mul_mat_q4_1_f32               0x15170fb20
ggml_metal_init: loaded kernel_mul_mat_q2_K_f32               0x151710380
ggml_metal_init: loaded kernel_mul_mat_q3_K_f32               0x151710a80
ggml_metal_init: loaded kernel_mul_mat_q4_K_f32               0x151711160
ggml_metal_init: loaded kernel_mul_mat_q5_K_f32               0x151711840
ggml_metal_init: loaded kernel_mul_mat_q6_K_f32               0x151712120
ggml_metal_init: loaded kernel_rope                           0x151712a10
ggml_metal_init: loaded kernel_alibi_f32                      0x1517132d0
ggml_metal_init: loaded kernel_cpy_f32_f16                    0x151713b60
ggml_metal_init: loaded kernel_cpy_f32_f32                    0x151714650
ggml_metal_init: loaded kernel_cpy_f16_f16                    0x151714dc0
ggml_metal_init: recommendedMaxWorkingSetSize = 21845.34 MB
ggml_metal_init: hasUnifiedMemory             = true
ggml_metal_init: maxTransferRate              = built-in GPU
llama_new_context_with_model: max tensor size =    87.89 MB
ggml_metal_add_buffer: allocated 'data            ' buffer, size =  6983.72 MB, ( 6984.17 / 21845.34)
ggml_metal_add_buffer: allocated 'eval            ' buffer, size =  1024.00 MB, ( 8008.17 / 21845.34)
ggml_metal_add_buffer: allocated 'kv              ' buffer, size =   402.00 MB, ( 8410.17 / 21845.34)
ggml_metal_add_buffer: allocated 'scr0            ' buffer, size =   512.00 MB, ( 8922.17 / 21845.34)
ggml_metal_add_buffer: allocated 'scr1            ' buffer, size =   512.00 MB, ( 9434.17 / 21845.34)

system_info: n_threads = 8 / 12 | AVX = 0 | AVX2 = 0 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 0 | NEON = 1 | ARM_FMA = 1 | F16C = 0 | FP16_VA = 1 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 0 | VSX = 0 |
sampling: repeat_last_n = 64, repeat_penalty = 1.100000, presence_penalty = 0.000000, frequency_penalty = 0.000000, top_k = 40, tfs_z = 1.000000, top_p = 0.950000, typical_p = 1.000000, temp = 0.800000, mirostat = 0, mirostat_lr = 0.100000, mirostat_ent = 5.000000
generate: n_ctx = 512, n_batch = 512, n_predict = 128, n_keep = 0



// RUN: %target-swift-ide-test -code-completion -source-filename %s -code-completion-token=GLOBAL_ENUM_1 > %t.globaleenum1.txt
// RUN: FileCheck --input-file=%t.globaleenum1.txt %s

// CHECK: CODE_ completionsForGlobalEnum1
// CHECK-NEXT: SYMBOL KIND
// CHECK-NEXT: BASE_ENUM=TypeName
// CHECK-NEXT: ◂────────────────────────
llama_print_timings:        load time =  3865.14 ms
llama_print_timings:      sample time =   182.48 ms /   128 runs   (    1.43 ms per token,   701.45 tokens per second)
llama_print_timings: prompt eval time =   352.93 ms /     2 tokens (  176.47 ms per token,     5.67 tokens per second)
llama_print_timings:        eval time =  4750.40 ms /   127 runs   (   37.40 ms per token,    26.73 tokens per second)
llama_print_timings:       total time =  5304.79 ms
ggml_metal_free: deallocating

17:54:24 in ~/Developer/ai/llama.cpp on  mps-q4_0-kernel   took 9.4s
➜ git co -
Switched to branch 'master'
Your branch is up to date with 'origin/master'.

17:54:31 in ~/Developer/ai/llama.cpp on  master 
➜ LLAMA_METAL=1 make
I llama.cpp build info:
I UNAME_S:  Darwin
I UNAME_P:  arm
I UNAME_M:  arm64
I CFLAGS:   -I.              -O3 -std=c11   -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wdouble-promotion -Wshadow -Wstrict-prototypes -Wpointer-arith -pthread -DGGML_USE_K_QUANTS -DGGML_USE_ACCELERATE -DGGML_USE_METAL -DGGML_METAL_NDEBUG
I CXXFLAGS: -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL
I LDFLAGS:   -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
I CC:       Apple clang version 14.0.3 (clang-1403.0.22.14.1)
I CXX:      Apple clang version 14.0.3 (clang-1403.0.22.14.1)

cc -I.              -O3 -std=c11   -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wdouble-promotion -Wshadow -Wstrict-prototypes -Wpointer-arith -pthread -DGGML_USE_K_QUANTS -DGGML_USE_ACCELERATE -DGGML_USE_METAL -DGGML_METAL_NDEBUG -c ggml-metal.m -o ggml-metal.o
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/main/main.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o main  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders

====  Run ./main -h for help.  ====

c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/quantize/quantize.cpp ggml.o llama.o k_quants.o ggml-metal.o -o quantize  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/quantize-stats/quantize-stats.cpp ggml.o llama.o k_quants.o ggml-metal.o -o quantize-stats  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/perplexity/perplexity.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o perplexity  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/embedding/embedding.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o embedding  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL pocs/vdot/vdot.cpp ggml.o k_quants.o ggml-metal.o -o vdot  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/train-text-from-scratch/train-text-from-scratch.cpp ggml.o llama.o k_quants.o ggml-metal.o -o train-text-from-scratch  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/simple/simple.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o simple  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL -Iexamples/server examples/server/server.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o server  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ --shared -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/embd-input/embd-input-lib.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o libembdinput.so  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders
c++ -I. -I./examples -O3 -std=c++11 -fPIC -DNDEBUG -Wall -Wextra -Wpedantic -Wcast-qual -Wno-unused-function -Wno-multichar -pthread -DGGML_USE_K_QUANTS -DGGML_USE_METAL examples/embd-input/embd-input-test.cpp ggml.o llama.o common.o k_quants.o ggml-metal.o -o embd-input-test  -framework Accelerate -framework Foundation -framework Metal -framework MetalKit -framework MetalPerformanceShaders -L. -lembdinput

17:54:44 in ~/Developer/ai/llama.cpp on  master   took 12.5s
➜ ./main -m ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin -n 128 -c 512 -s 12 -ngl 1 --no-mmap -p ""
main: build = 823 (4e7464e)
main: seed  = 12
llama.cpp: loading model from ../models/TheBloke_WizardLM-13B-V1.1-GGML/wizardlm-13b-v1.1.ggmlv3.q4_0.bin
llama_model_load_internal: format     = ggjt v3 (latest)
llama_model_load_internal: n_vocab    = 32001
llama_model_load_internal: n_ctx      = 512
llama_model_load_internal: n_embd     = 5120
llama_model_load_internal: n_mult     = 256
llama_model_load_internal: n_head     = 40
llama_model_load_internal: n_layer    = 40
llama_model_load_internal: n_rot      = 128
llama_model_load_internal: ftype      = 2 (mostly Q4_0)
llama_model_load_internal: n_ff       = 13824
llama_model_load_internal: model size = 13B
llama_model_load_internal: ggml ctx size = 6983.71 MB
llama_model_load_internal: mem required  = 9031.71 MB (+ 1608.00 MB per state)
llama_new_context_with_model: kv self size  =  400.00 MB
ggml_metal_init: allocating
ggml_metal_init: using MPS
ggml_metal_init: loading '/Volumes/Developer/ai/llama.cpp/ggml-metal.metal'
ggml_metal_init: loaded kernel_add                            0x1236044a0
ggml_metal_init: loaded kernel_mul                            0x123604ce0
ggml_metal_init: loaded kernel_mul_row                        0x123605310
ggml_metal_init: loaded kernel_scale                          0x123605830
ggml_metal_init: loaded kernel_silu                           0x123605d50
ggml_metal_init: loaded kernel_relu                           0x123606270
ggml_metal_init: loaded kernel_gelu                           0x123606790
ggml_metal_init: loaded kernel_soft_max                       0x123606e40
ggml_metal_init: loaded kernel_diag_mask_inf                  0x1236074a0
ggml_metal_init: loaded kernel_get_rows_f16                   0x123607b20
ggml_metal_init: loaded kernel_get_rows_q4_0                  0x1236081a0
ggml_metal_init: loaded kernel_get_rows_q4_1                  0x123608990
ggml_metal_init: loaded kernel_get_rows_q2_K                  0x144106970
ggml_metal_init: loaded kernel_get_rows_q3_K                  0x144107110
ggml_metal_init: loaded kernel_get_rows_q4_K                  0x144107790
ggml_metal_init: loaded kernel_get_rows_q5_K                  0x144107e10
ggml_metal_init: loaded kernel_get_rows_q6_K                  0x144108490
ggml_metal_init: loaded kernel_rms_norm                       0x144108b40
…